+++ scripts/build-instrumentation-tests.sh
@@ -1,2 +1,0 @@
-# Build the Espresso Test App
-./gradlew assembleAndroidTest --stacktrace